move get_score to base
