+++ sdmetrics/reports/single_table/quality_report.py
@@ -16,13 +16,3 @@ class QualityReport(BaseReport):
             'Column Shapes': ColumnShapes(),
             'Column Pair Trends': ColumnPairTrends()
         }
-
-    def get_score(self):
-        """Return the overall quality score.
-
-        Returns:
-            float
-                The overall quality score.
-        """
-        self._check_report_generated()
-        return self._overall_score
