+++ tests/integration/reports/single_table/test_quality_report.py
@@ -1,6 +1,3 @@
-import contextlib
-import io
-import re
 import time
 from datetime import date, datetime
