+++ tests/unit/reports/single_table/test_quality_report.py
@@ -1,5 +1,3 @@
-from unittest.mock import Mock, patch
-
 from sdmetrics.reports.single_table import QualityReport
 from sdmetrics.reports.single_table._properties import ColumnPairTrends, ColumnShapes
 
@@ -16,17 +14,3 @@ class TestQualityReport:
         assert not report.is_generated
         assert isinstance(report._properties['Column Shapes'], ColumnShapes)
         assert isinstance(report._properties['Column Pair Trends'], ColumnPairTrends)
-
-    def test_get_score(self):
-        """Test the ``get_score`` method."""
-        # Setup
-        quality_report = QualityReport()
-        quality_report.is_generated = True
-        mock_score = Mock()
-        quality_report._overall_score = mock_score
-
-        # Run
-        score = quality_report.get_score()
-
-        # Assert
-        assert score == mock_score
